修改

+++ src/views/Home.vue
@@ -3,8 +3,8 @@
     <!-- 头部 -->
     <el-header class="hd">
       <div class="title">
-        <img src="../assets/logo.png" alt />
-        <p>源码漏洞扫描系统</p>
+        <img src="../images/logo1.gif" alt />
+        <p @click="jump">源码漏洞扫描系统</p>
       </div>
       <el-button type="success" class="back" v-show="isFlag" @click="goback">返回项目列表</el-button>
       <el-button type="primary" class="logout" @click="exit">退出</el-button>
@@ -33,15 +33,30 @@ export default {
 		}
 	},
 	methods: {
-		exit() {
-			//退出登录就要清除sessionStorage中的token 并且跳转登录页面
-			window.sessionStorage.clear()
-			this.$router.push('/login')
+		async exit() {
+			let res = await this.$confirm('确定退出系统？', '提示', {
+				confirmButtonText: '确定',
+				cancelButtonText: '取消',
+				type: 'warning'
+			}).catch(() => {
+				//catch方法表示此操作失败或者取消的处理函数
+				
+			})
+			if (res == 'confirm') {
+				//confirm确认删除
+				//退出登录就要清除sessionStorage中的token 并且跳转登录页面
+				window.sessionStorage.clear()
+				this.$router.push('/login')
+			}
 		},
 		goback() {
 			this.$router.push('list')
 			this.$store.commit('setMark', false)
-		}
+    },
+    jump(){
+      this.$router.push('/list')
+      this.$store.commit('setMark', false)
+    }
 	}
 }
 </script>
@@ -52,7 +67,7 @@ export default {
 }
 .el-header {
 	background-color: #0a2438;
-	height: unset !important;
+	height: 44px !important;
 	// display: flex;
 }
 .el-main {
@@ -72,16 +87,21 @@ export default {
 	padding: 8px 10px;
 }
 .title {
+  height: 100%;
 	display: flex;
 	align-items: center;
 	img {
-		height: 46px;
+    height: 36px;
+    border-radius: 50%;
 	}
 	p {
-    margin-left: 20px;
-    margin-top: 8px;
-    margin-bottom: 8px;
-		color: #fff;
+		margin-left: 20px;
+		margin-top: 8px;
+		margin-bottom: 8px;
+    color: #fff;
+    font-size: 16px;
+    font-weight: 700;
+    cursor: pointer;
 	}
 }
 </style>
